start emitting responses over pipes
